# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorMessage {
  error: string;
}

export interface Attendee {
  id: number;
  firstName: string;
  lastName: string;
  city: string;
  state: string;
  mobileNumber: string;
  mobileVerified: boolean;
  businessType: string;
  createdAt: string;
}

export interface AttendeeInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 7 */
  mobileNumber: string;
  /** @minLength 1 */
  businessType: string;
}

export interface VerificationInput {
  /**
     * @minLength 6
     * @maxLength 6
     */
  code: string;
}

export type StoryResponseType = typeof StoryResponseType[keyof typeof StoryResponseType];


export const StoryResponseType = {
  voice: 'voice',
  written: 'written',
} as const;

export interface Story {
  id: number;
  attendeeId: number;
  responseType: StoryResponseType;
  /** @nullable */
  writtenResponse?: string | null;
  /** @nullable */
  voiceDurationSeconds?: number | null;
  /** @nullable */
  promptShown?: string | null;
  /** @nullable */
  selectedInterests?: string[] | null;
  /** @nullable */
  followUpPrompt?: string | null;
  /** @nullable */
  followUpResponse?: string | null;
  /** @nullable */
  sentiment?: string | null;
  /** @nullable */
  detectedTheme?: string | null;
  createdAt: string;
}

export type StoryInputResponseType = typeof StoryInputResponseType[keyof typeof StoryInputResponseType];


export const StoryInputResponseType = {
  voice: 'voice',
  written: 'written',
} as const;

export interface StoryInput {
  attendeeId: number;
  responseType: StoryInputResponseType;
  writtenResponse?: string;
  voiceDurationSeconds?: number;
  promptShown?: string;
  selectedInterests?: string[];
}

export interface FollowUpInput {
  /** @minLength 1 */
  followUpPrompt: string;
  followUpResponse?: string;
}

export interface BusinessTypeShare {
  businessType: string;
  percent: number;
}

export interface AttendeeVoice {
  quote: string;
  businessType: string;
}

export interface IntelligenceSummary {
  storiesShared: number;
  voicePercent: number;
  writtenPercent: number;
  topBusinessTypes: BusinessTypeShare[];
  topThemes: string[];
  emergingStoryHeadline: string;
  emergingStoryInsight: string;
  attendeeVoices: AttendeeVoice[];
}