# lib/api-zod/src/generated/types/attendee.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Attendee {
  id: number;
  firstName: string;
  lastName: string;
  city: string;
  state: string;
  mobileNumber: string;
  mobileVerified: boolean;
  businessType: string;
  createdAt: string;
}