# lib/api-zod/src/generated/types/intelligenceSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendeeVoice } from './attendeeVoice';
import type { BusinessTypeShare } from './businessTypeShare';

export interface IntelligenceSummary {
  storiesShared: number;
  voicePercent: number;
  writtenPercent: number;
  topBusinessTypes: BusinessTypeShare[];
  topThemes: string[];
  emergingStoryHeadline: string;
  emergingStoryInsight: string;
  attendeeVoices: AttendeeVoice[];
}