# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoryResponseType } from './storyResponseType';

export interface Story {
  id: number;
  attendeeId: number;
  responseType: StoryResponseType;
  /** @nullable */
  writtenResponse?: string | null;
  /** @nullable */
  voiceDurationSeconds?: number | null;
  /** @nullable */
  promptShown?: string | null;
  /** @nullable */
  selectedInterests?: string[] | null;
  /** @nullable */
  followUpPrompt?: string | null;
  /** @nullable */
  followUpResponse?: string | null;
  /** @nullable */
  sentiment?: string | null;
  /** @nullable */
  detectedTheme?: string | null;
  createdAt: string;
}