# lib/api-zod/src/generated/types/storyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoryInputResponseType } from './storyInputResponseType';

export interface StoryInput {
  attendeeId: number;
  responseType: StoryInputResponseType;
  writtenResponse?: string;
  voiceDurationSeconds?: number;
  promptShown?: string;
  selectedInterests?: string[];
}