# lib/api-zod/src/generated/types/verificationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationInput {
  /**
     * @minLength 6
     * @maxLength 6
     */
  code: string;
}